Fixed drug quantities to be decimals rather than integers.
CBBD-232

+++ bluebutton-data-pipeline-rif/src/main/java/gov/hhs/cms/bluebutton/datapipeline/rif/model/HospiceClaimGroup.java
@@ -340,7 +340,7 @@ public final class HospiceClaimGroup {
 		/**
 		 * @see Column#REV_CNTR_NDC_QTY
 		 */
-		public Optional<Integer> nationalDrugCodeQuantity;
+		public Optional<BigDecimal> nationalDrugCodeQuantity;
 
 		/**
 		 * @see Column#REV_CNTR_NDC_QTY_QLFR_CD
